+++ Demo/Demo/DemoApp.swift
@@ -3,6 +3,7 @@
 //  Demo
 //
 //  Created by Daniel Saidi on 2024-11-23.
+//  Copyright © 2024-2026 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI