+++ Sources/AppIconKit/AlternateAppIconCollection.swift
@@ -3,15 +3,14 @@
 //  AppIconKit
 //
 //  Created by Daniel Saidi on 2024-11-22.
-//  Copyright © 2024-2025 Daniel Saidi. All rights reserved.
+//  Copyright © 2024-2026 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
 
-/// This collection can be used to group alternate app icons
-/// into related groups.
+/// This collection can be used to group alternate app icons into related groups.
 public struct AlternateAppIconCollection {
-    
+
     /// Create an alternate app icon collection.
     ///
     /// - Parameters:
@@ -24,20 +23,18 @@ public struct AlternateAppIconCollection {
         self.name = name
         self.icons = icons
     }
-    
-    /// Create an alternate app icon collection by mapping a
-    /// set of icon names to ``AlternateAppIcon`` values.
+
+    /// Create an icon collection by mapping a set of icon names to icon values.
     ///
-    /// The `appIconName` should be the name of the main app
-    /// icon `.appiconset` asset that is used by default.
+    /// The ``appIconName`` should be the default app icon's asset name.
     ///
-    /// The `iconNames` array should be a list of plain icon
-    /// names, while `iconNamePrefix` is an icon name prefix
-    /// that will be added to the `.imageset` asset, and the
-    /// `appIconNamePrefix` to the `.appiconset` asset.
+    /// The `iconNames` list should be a list of plain icon names, where each
+    /// icon name will be prefixed by the `iconNamePrefix` when resolving
+    /// the name of the `.imageset` asset, and the `appIconNamePrefix`
+    /// when resolving the name of the `.appiconset` asset.
     ///
-    /// If the resulting icon name equals `appIconName`, the
-    /// icon will reset the alternate app icon when selected.
+    /// If the resulting icon name equals the default `appIconName`, the icon
+    /// will reset the alternate app icon when it's selected.
     ///
     /// - Parameters:
     ///   - name: The name of the section.
@@ -60,6 +57,7 @@ public struct AlternateAppIconCollection {
                 let isDefaultIcon = imageName == appIconName
                 let appIconName = isDefaultIcon ? nil : "\(appIconNamePrefix)\(name)"
                 return .init(
+                    name: imageName,
                     icon: Image(imageName),
                     appIconName: appIconName
                 )