+++ Sources/AppIconKit/AlternateAppIconContext.swift
@@ -3,14 +3,13 @@
 //  AppIconKit
 //
 //  Created by Daniel Saidi on 2024-11-22.
-//  Copyright © 2024-2025 Daniel Saidi. All rights reserved.
+//  Copyright © 2024-2026 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
 import SwiftUI
 
-/// This observable context class can be used to managed the
-/// alternate app icon for an iOS or macOS app.
+/// This observable context class can be used to managed the alternate app icon.
 @MainActor
 open class AlternateAppIconContext: NSObject, ObservableObject {